backend/requirements.txt
# Stub scripts (throwaway)
pymongo[srv]>=4.6.0
openai>=1.0.0
python-dotenv>=1.0.0

# Backend
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.3.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
structlog>=24.1.0
boto3>=1.34.0
open-clip-torch>=2.24.0
google-genai>=1.0.0
Pillow>=10.3.0

# Tests
pytest>=8.0.0
httpx>=0.27.0
